+++ R/correct_attributes.R
@@ -2,21 +2,21 @@
 #'
 #' SPSS users frequently label their missing values, but don't set them as missing.
 #' This function will rectify that for negative values and for the values 99 and 999 (only if they're 5*MAD away from the median).
-#' Using different settings, you can also easily tag other missings.
+#' Using different settings, you can also easily tag other missing values.
 #'
-#' @param data the data frame with labelled missings
-#' @param only_labelled_missings don't set values to missing if there's no label for them
+#' @param data the data frame with labelled missing values
+#' @param only_labelled don't set values to missing if there's no label for them
 #' @param negative_values_are_missing by default we label negative values as missing
 #' @param ninety_nine_problems SPSS users often store values as 99/999, should we do this for values with 5*MAD of the median
-#' @param learn_from_labels if there are labels for missings of the form `[-1] no answer`, set -1 in the data to the corresponding tagged missing
+#' @param learn_from_labels if there are labels for missing values of the form `[-1] no answer`, set -1 in the data to the corresponding tagged missing
 #' @param missing also set these values to missing (or enforce for 99/999 within 5*MAD)
 #' @param non_missing don't set these values to missing
 #' @param vars only edit these variables
-#' @param use_labelled_spss the labelled_spss class has a few drawbacks. Since R can't store missings like -1 and 99, we're replacing them with letters unless this option is enabled. If you prefer to keep your -1 etc, turn this on.
+#' @param use_labelled_spss the labelled_spss class has a few drawbacks. Since R can't store missing values like -1 and 99, we're replacing them with letters unless this option is enabled. If you prefer to keep your -1 etc, turn this on.
 #'
 #' @export
 #'
-detect_missings <- function(data, only_labelled_missings = TRUE,
+detect_missing <- function(data, only_labelled = TRUE,
                                     negative_values_are_missing = TRUE,
                                     ninety_nine_problems = TRUE,
                                     learn_from_labels = TRUE,
@@ -28,9 +28,9 @@ detect_missings <- function(data, only_labelled_missings = TRUE,
     if (is.numeric(data[[ var ]]) && any(!is.na(data[[ var ]]))) {
 
       # negative values
-      potential_missings <- c()
+      potential_missing_values <- c()
       if (negative_values_are_missing) {
-        potential_missings <- unique(data[[var]][data[[var]] < 0])
+        potential_missing_values <- unique(data[[var]][data[[var]] < 0])
       }
       labels <- attributes(data[[var]])$labels
 
@@ -41,8 +41,8 @@ detect_missings <- function(data, only_labelled_missings = TRUE,
         potential_tags <- labels[is.na(labels)]
         if (!all(is.na(haven::na_tag(data[[var]]))) &&
             length(intersect(potentially_untagged, data[[var]]))) {
-          warning("There were already tagged missings in ", var, ". Although",
-                  "there were further potential missings as indicated by",
+          warning("Missing values were already tagged in ", var, ". Although",
+                  "there were further potential missing values as indicated by",
                   "missing labels, this was not changed.")
         } else {
           for (e in seq_along(potentially_untagged)) {
@@ -51,44 +51,44 @@ detect_missings <- function(data, only_labelled_missings = TRUE,
           }
         }
       }
-      # classic SPSS missings only if they are far out of real data range
+      # classic SPSS missing values only if they are far out of real data range
       # can be turned off using non_missing or ninety_nine_problems
       if (ninety_nine_problems) {
         if (any(!is.na(data[[ var ]])) &&
             (stats::median(data[[var]], na.rm = TRUE) +
              stats::mad(data[[var]], na.rm = TRUE) * 5) < 99) {
-          potential_missings <- c(potential_missings, 99)
+          potential_missing_values <- c(potential_missing_values, 99)
         }
         if (any(!is.na(data[[ var ]])) &&
             (stats::median(data[[var]], na.rm = TRUE) +
              stats::mad(data[[var]], na.rm = TRUE) * 5) < 999) {
-          potential_missings <- c(potential_missings, 999)
+          potential_missing_values <- c(potential_missing_values, 999)
         }
       }
-      potential_missings <- union(
-        setdiff(potential_missings, non_missing),
+      potential_missing_values <- union(
+        setdiff(potential_missing_values, non_missing),
         missing)
-      if ((!only_labelled_missings || haven::is.labelled(data[[var]])) &&
-          length(potential_missings) > 0) {
-        if (only_labelled_missings) {
-          potential_missings <- potential_missings[
-            potential_missings %in% labels]
-          # add labelled missings that don't exist for completeness
-          potential_missings <- union(potential_missings,
+      if ((!only_labelled || haven::is.labelled(data[[var]])) &&
+          length(potential_missing_values) > 0) {
+        if (only_labelled) {
+          potential_missing_values <- potential_missing_values[
+            potential_missing_values %in% labels]
+          # add labelled missing_values that don't exist for completeness
+          potential_missing_values <- union(potential_missing_values,
             setdiff(labels, data[[var]]))
         }
-        potential_missings <- sort(potential_missings)
+        potential_missing_values <- sort(potential_missing_values)
         with_tagged_na <- data[[var]]
         free_na_tags <- setdiff( letters, haven::na_tag(with_tagged_na))
 
-        for (i in seq_along(potential_missings)) {
-          miss <- potential_missings[i]
+        for (i in seq_along(potential_missing_values)) {
+          miss <- potential_missing_values[i]
 
           if (!use_labelled_spss &&
-              !all(potential_missings %in% free_na_tags)) {
+              !all(potential_missing_values %in% free_na_tags)) {
             new_miss <- free_na_tags[i]
           } else {
-            new_miss <- potential_missings[i]
+            new_miss <- potential_missing_values[i]
           }
           that_label <- which(labels == miss)
           if (!use_labelled_spss) {
@@ -97,19 +97,20 @@ detect_missings <- function(data, only_labelled_missings = TRUE,
           }
           if (length(that_label) && !use_labelled_spss) {
             labels[that_label] <- haven::tagged_na(new_miss)
-            names(labels)[that_label] <- paste0("[", potential_missings[i],
+            names(labels)[that_label] <- paste0("[",
+                                  potential_missing_values[i],
                                       "] ", names(labels)[that_label])
           }
         }
         if (use_labelled_spss) {
           labels <- attributes(data[[var]])$labels
           if (is.null(labels)) {
-            labels <- potential_missings
+            labels <- potential_missing_values
             names(labels) <- "autodetected unlabelled missing"
           }
           data[[var]] <- haven::labelled_spss(data[[var]],
                                  labels = labels,
-                                 na_values = potential_missings,
+                                 na_values = potential_missing_values,
                                  na_range = attributes(data[[var]])$na_range)
         } else if (haven::is.labelled(data[[var]])) {
             data[[var]] <- haven::labelled(with_tagged_na, labels = labels)
@@ -120,6 +121,18 @@ detect_missings <- function(data, only_labelled_missings = TRUE,
     }
   }
   data
+}
+#' @describeIn detect_missing Deprecated version
+#' @param only_labelled_missings passed to [detect_missing()]
+#' @param ... passed to [detect_missing()]
+#' @inheritParams detect_missing
+#' @export
+detect_missings <- function(data, only_labelled_missings = TRUE,
+                           ...) {
+  .Deprecated("detect_missing", package = "codebook", msg =
+                "We renamed this function so as not to incorrectly pluralise
+the word 'missing'.")
+  detect_missing(data, only_labelled = only_labelled_missings, ...)
 }
 
 #' Rescue lost attributes